Add interactive page
