+++ modules/Elements.js
@@ -1,0 +1,3 @@
+export const list = document.querySelector('ul');
+export const form = document.querySelector('form');
+export const clear = document.querySelector('.clear-btn');
