+++ modules/GetTodo.js
@@ -1,0 +1,13 @@
+import data from './todoData.js';
+import display from './Display.js';
+
+const getTodos = () => {
+  document.addEventListener('DOMContentLoaded', () => {
+    if (localStorage.getItem('todos')) {
+      data.todos = JSON.parse(localStorage.getItem('todos'));
+      display();
+    }
+  });
+};
+
+export default getTodos;
